Added generic component for articles and tutorials page content
