+++ src/pages/articles.jsx
@@ -1,106 +1,11 @@
 /* eslint-disable react/prop-types */
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { graphql } from 'gatsby';
 import { navigate } from '@reach/router';
 
 import Layout from '../components/layout';
-import ArticlePreviewList from '../components/ArticlePreviewList';
-import HeaderSearchArea from '../components/HeaderSearchArea';
-import Pager from '../components/Pager';
+import ContentPageItems from '../components/ContentPageItems';
 
-import utils from '../utils';
-
-import {
-  contentPage,
-  contentSection,
-  pager,
-} from '../styles/contentPage.module.scss';
-
-const ArticleListContainer = ({ articleEdges, initCurrentPage }) => {
-  const [searchValue, setSearchValue] = useState('');
-  const [selectedTags, setSelectedTags] = useState(['ALL']);
-  const [currentPage, setCurrentPage] = useState(initCurrentPage);
-
-  useEffect(() => {
-    setCurrentPage(initCurrentPage);
-  }, [initCurrentPage]);
-
-  const itemsPerPage = 9;
-
-  const handleSearchValueChange = (event) => setSearchValue(event.target.value);
-
-  const handleTagToggle = (tagValue) => {
-    const newSelectedTags = utils.getNewTagArray(tagValue, selectedTags);
-    setSelectedTags([...newSelectedTags]);
-  };
-
-  const handlePageChange = (nextPageNum) => {
-    setCurrentPage(nextPageNum);
-    navigate(`/articles?page=${nextPageNum || currentPage}`);
-  };
-
-  const filteredArticles = utils.filterItems(
-    articleEdges,
-    searchValue,
-    selectedTags
-  );
-
-  const currentPageArticles = utils.getItemsForPage(
-    currentPage,
-    itemsPerPage,
-    filteredArticles,
-    handlePageChange,
-    false
-  );
-
-  // Array of the page numbers required for the pager component
-  const pageArray = utils.getPages(filteredArticles, itemsPerPage);
-
-  return (
-    <div className={contentPage}>
-      <HeaderSearchArea
-        title="Articles on Developer Experience"
-        description="Thoughts and experiences from experienced practicioners in the field"
-        searchPlaceholder="Search Articles"
-        searchValue={searchValue}
-        selectedTags={selectedTags}
-        tags={[
-          {
-            name: 'ALL',
-          },
-          {
-            name: 'DEVELOP',
-          },
-          {
-            name: 'MANAGE',
-          },
-          {
-            name: 'RESEARCH',
-          },
-        ]}
-        handleTagToggle={handleTagToggle}
-        handleSearchChange={handleSearchValueChange}
-      />
-      <div className={contentSection}>
-        {currentPageArticles.length > 0 ? (
-          <ArticlePreviewList
-            previewData={currentPageArticles}
-            type="articles"
-          />
-        ) : (
-          <p>No articles found</p>
-        )}
-      </div>
-      <div className={pager}>
-        <Pager
-          pages={pageArray}
-          currentPage={currentPage}
-          handleClick={handlePageChange}
-        />
-      </div>
-    </div>
-  );
-};
 
 const ArticlesPage = ({ data, location }) => {
   const articleEdges = data.allMarkdownRemark.edges;
@@ -115,8 +20,11 @@ const ArticlesPage = ({ data, location }) => {
 
   return (
     <Layout>
-      <ArticleListContainer
-        articleEdges={articleEdges}
+      <ContentPageItems
+        title="Articles on Developer Experience"
+        description="Thoughts and experiences from experienced practicioners in the field"
+        type="Articles"
+        itemEdges={articleEdges}
         initCurrentPage={currentPage}
       />
     </Layout>
